backend/services/auth-service/requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==1.10.14
redis>=5.0.0
python-dotenv==1.0.1
pytest==7.4.4
httpx==0.25.2
email-validator==2.1.0.post1
pydantic[email]==1.10.14
cryptography==42.0.5
